Generalize osmRelation and osmChangeset
