+++ modules/osm/relation.js
@@ -2,8 +2,9 @@ import { geoArea as d3_geoArea } from 'd3-geo';
 
 import { osmEntity } from './entity';
 import { entityEntity } from '../entities/entity';
+import { entityRelation } from '../entities/relation';
 import { osmJoinWays } from './multipolygon';
-import { geoExtent, geoPolygonContainsPolygon, geoPolygonIntersectsPolygon } from '../geo';
+import { geoPolygonContainsPolygon, geoPolygonIntersectsPolygon } from '../geo';
 
 
 export function osmRelation() {
@@ -17,245 +18,18 @@ export function osmRelation() {
 
 entityEntity.relation = osmRelation;
 
-osmRelation.prototype = Object.create(osmEntity.prototype);
+// inhert from entityRelation
+osmRelation.prototype = Object.create(entityRelation.prototype);
 
-
-osmRelation.creationOrder = function(a, b) {
-    var aId = parseInt(entityEntity.id.toUntyped(a.id), 10);
-    var bId = parseInt(entityEntity.id.toUntyped(b.id), 10);
-
-    if (aId < 0 || bId < 0) return aId - bId;
-    return bId - aId;
-};
-
+// Use common properties of osmEntity.
+// This is not prototype inheritance. osmRelation is not an `instanceof` osmEntity.
+Object.assign(osmRelation.prototype, osmEntity.prototype);
 
 Object.assign(osmRelation.prototype, {
-    type: 'relation',
-    members: [],
-
-
-    copy: function(resolver, copies) {
-        if (copies[this.id]) return copies[this.id];
-
-        var copy = osmEntity.prototype.copy.call(this, resolver, copies);
-
-        var members = this.members.map(function(member) {
-            return Object.assign({}, member, { id: resolver.entity(member.id).copy(resolver, copies).id });
-        });
-
-        copy = copy.update({members: members});
-        copies[this.id] = copy;
-
-        return copy;
-    },
-
-
-    extent: function(resolver, memo) {
-        return resolver.transient(this, 'extent', function() {
-            if (memo && memo[this.id]) return geoExtent();
-            memo = memo || {};
-            memo[this.id] = true;
-
-            var extent = geoExtent();
-            for (var i = 0; i < this.members.length; i++) {
-                var member = resolver.hasEntity(this.members[i].id);
-                if (member) {
-                    extent._extend(member.extent(resolver, memo));
-                }
-            }
-            return extent;
-        });
-    },
-
-
-    geometry: function(graph) {
-        return graph.transient(this, 'geometry', function() {
-            return this.isMultipolygon() ? 'area' : 'relation';
-        });
-    },
-
-
-    isDegenerate: function() {
-        return this.members.length === 0;
-    },
-
-
-    // Return an array of members, each extended with an 'index' property whose value
-    // is the member index.
-    indexedMembers: function() {
-        var result = new Array(this.members.length);
-        for (var i = 0; i < this.members.length; i++) {
-            result[i] = Object.assign({}, this.members[i], {index: i});
-        }
-        return result;
-    },
-
-
-    // Return the first member with the given role. A copy of the member object
-    // is returned, extended with an 'index' property whose value is the member index.
-    memberByRole: function(role) {
-        for (var i = 0; i < this.members.length; i++) {
-            if (this.members[i].role === role) {
-                return Object.assign({}, this.members[i], {index: i});
-            }
-        }
-    },
-
-    // Same as memberByRole, but returns all members with the given role
-    membersByRole: function(role) {
-        var result = [];
-        for (var i = 0; i < this.members.length; i++) {
-            if (this.members[i].role === role) {
-                result.push(Object.assign({}, this.members[i], {index: i}));
-            }
-        }
-        return result;
-    },
-
-    // Return the first member with the given id. A copy of the member object
-    // is returned, extended with an 'index' property whose value is the member index.
-    memberById: function(id) {
-        for (var i = 0; i < this.members.length; i++) {
-            if (this.members[i].id === id) {
-                return Object.assign({}, this.members[i], {index: i});
-            }
-        }
-    },
-
-
-    // Return the first member with the given id and role. A copy of the member object
-    // is returned, extended with an 'index' property whose value is the member index.
-    memberByIdAndRole: function(id, role) {
-        for (var i = 0; i < this.members.length; i++) {
-            if (this.members[i].id === id && this.members[i].role === role) {
-                return Object.assign({}, this.members[i], {index: i});
-            }
-        }
-    },
-
-
-    addMember: function(member, index) {
-        var members = this.members.slice();
-        members.splice(index === undefined ? members.length : index, 0, member);
-        return this.update({members: members});
-    },
-
-
-    updateMember: function(member, index) {
-        var members = this.members.slice();
-        members.splice(index, 1, Object.assign({}, members[index], member));
-        return this.update({members: members});
-    },
-
-
-    removeMember: function(index) {
-        var members = this.members.slice();
-        members.splice(index, 1);
-        return this.update({members: members});
-    },
-
-
-    removeMembersWithID: function(id) {
-        var members = this.members.filter(function(m) { return m.id !== id; });
-        return this.update({members: members});
-    },
-
-    moveMember: function(fromIndex, toIndex) {
-        var members = this.members.slice();
-        members.splice(toIndex, 0, members.splice(fromIndex, 1)[0]);
-        return this.update({members: members});
-    },
-
-
-    // Wherever a member appears with id `needle.id`, replace it with a member
-    // with id `replacement.id`, type `replacement.type`, and the original role,
-    // By default, adding a duplicate member (by id and role) is prevented.
-    // Return an updated relation.
-    replaceMember: function(needle, replacement, keepDuplicates) {
-        if (!this.memberById(needle.id)) return this;
-
-        var members = [];
-
-        for (var i = 0; i < this.members.length; i++) {
-            var member = this.members[i];
-            if (member.id !== needle.id) {
-                members.push(member);
-            } else if (keepDuplicates || !this.memberByIdAndRole(replacement.id, member.role)) {
-                members.push({ id: replacement.id, type: replacement.type, role: member.role });
-            }
-        }
-
-        return this.update({ members: members });
-    },
-
-
-    asJXON: function(changeset_id) {
-        var r = {
-            relation: {
-                '@id': this.untypedID(),
-                '@version': this.version || 0,
-                member: this.members.map(function(member) {
-                    return {
-                        keyAttributes: {
-                            type: member.type,
-                            role: member.role,
-                            ref: entityEntity.id.toUntyped(member.id)
-                        }
-                    };
-                }, this),
-                tag: Object.keys(this.tags).map(function(k) {
-                    return { keyAttributes: { k: k, v: this.tags[k] } };
-                }, this)
-            }
-        };
-        if (changeset_id) {
-            r.relation['@changeset'] = changeset_id;
-        }
-        return r;
-    },
-
-
-    asGeoJSON: function(resolver) {
-        return resolver.transient(this, 'GeoJSON', function () {
-            if (this.isMultipolygon()) {
-                return {
-                    type: 'MultiPolygon',
-                    coordinates: this.multipolygon(resolver)
-                };
-            } else {
-                return {
-                    type: 'FeatureCollection',
-                    properties: this.tags,
-                    features: this.members.map(function (member) {
-                        return Object.assign({role: member.role}, resolver.entity(member.id).asGeoJSON(resolver));
-                    })
-                };
-            }
-        });
-    },
-
-
-    area: function(resolver) {
-        return resolver.transient(this, 'area', function() {
-            return d3_geoArea(this.asGeoJSON(resolver));
-        });
-    },
-
 
     isMultipolygon: function() {
         return this.tags.type === 'multipolygon';
     },
-
-
-    isComplete: function(resolver) {
-        for (var i = 0; i < this.members.length; i++) {
-            if (!resolver.hasEntity(this.members[i].id)) {
-                return false;
-            }
-        }
-        return true;
-    },
-
 
     hasFromViaTo: function() {
         return (
@@ -265,11 +39,9 @@ Object.assign(osmRelation.prototype, {
         );
     },
 
-
     isRestriction: function() {
         return !!(this.tags.type && this.tags.type.match(/^restriction:?/));
     },
-
 
     isValidRestriction: function() {
         if (!this.isRestriction()) return false;
@@ -289,7 +61,6 @@ Object.assign(osmRelation.prototype, {
 
         return true;
     },
-
 
     // Returns an array [A0, ... An], each Ai being an array of node arrays [Nds0, ... Ndsm],
     // where Nds0 is an outer ring and subsequent Ndsi's (if any i > 0) being inner rings.
@@ -354,4 +125,5 @@ Object.assign(osmRelation.prototype, {
 
         return result;
     }
+
 });
