+++ modules/ui/feature_list.js
@@ -10,7 +10,6 @@ import { coreGraph } from '../core/graph';
 import { geoSphericalDistance } from '../geo/geo';
 import { geoExtent, geoChooseEdge } from '../geo';
 import { modeSelect } from '../modes/select';
-import { osmEntity } from '../osm/entity';
 import { entityEntity } from '../entities/entity';
 import { services } from '../services';
 import { svgIcon } from '../svg/icon';
@@ -189,7 +188,7 @@ export function uiFeatureList(context) {
                     attrs.nodes = ['a','a'];  // so that geometry area is possible
                 }
 
-                var tempEntity = osmEntity(attrs);
+                var tempEntity = entityEntity(attrs);
                 var tempGraph = coreGraph([tempEntity]);
                 var matched = context.presets().match(tempEntity, tempGraph);
                 var type = (matched && matched.name()) || utilDisplayType(id);
